backend/requirements.txt
Django==2.2.19
djangorestframework==3.13.1
drf-extra-fields==3.4.0
django-filter==21.1
djoser==2.1.0
pytz==2022.2.1
sqlparse==0.4.2
python-dotenv==0.20.0
Pillow==9.2.0
reportlab==3.6.11
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6